+++ packages/style-system/src/components/base.ts
@@ -46,8 +46,16 @@ export const createFlexChildCSSProps = ({grow, shrink, basis, order, align}: Fle
   return styles
 }
 
+function getShadow<T extends BaseTheme>(theme: T, shadow: string): string {
+  if (shadow in theme.shadows) {
+    return theme.shadows[shadow]
+  }
+  return 'none'
+}
+
 export function createStylingCSSProps<T extends BaseTheme>(
-  {bg, cursorPointer, overflowHidden, elevation}: StylingProps<T> & WithBaseTheme, theme: BaseTheme) {
+  {bg, cursorPointer, overflowHidden, shadow, r, rt, rb, rl, rr, rtl, rtr, rbl, rbr}: StylingProps<T> & WithBaseTheme,
+  theme: BaseTheme) {
   const styles: IStyle = {}
   if (bg) {
     Object.assign(styles, {backgroundColor: getColor(theme, bg as string)})
@@ -58,8 +66,39 @@ export function createStylingCSSProps<T extends BaseTheme>(
   if (overflowHidden) {
     Object.assign(styles, {overflow: 'hidden'})
   }
-  if (elevation) {
-    Object.assign(styles, {boxShadow: lookupElevation(theme, elevation)})
+  if (shadow && shadow in theme.shadows) {
+    Object.assign(styles, {boxShadow: getShadow(theme, shadow as string)})
+  }
+  if (r) {
+    Object.assign(styles, {borderRadius: r})
+  }
+  if (rt) {
+    Object.assign(styles, {borderTopLeftRadius: rt})
+    Object.assign(styles, {borderTopRightRadius: rt})
+  }
+  if (rb) {
+    Object.assign(styles, {borderBottomRightRadius: rb})
+    Object.assign(styles, {borderBottomLeftRadius: rb})
+  }
+  if (rl) {
+    Object.assign(styles, {borderTopLeftRadius: rl})
+    Object.assign(styles, {borderBottomLeftRadius: rl})
+  }
+  if (rr) {
+    Object.assign(styles, {borderBottomRightRadius: rr})
+    Object.assign(styles, {borderTopRightRadius: rr})
+  }
+  if (rtl) {
+    Object.assign(styles, {borderTopLeftRadius: rtl})
+  }
+  if (rtr) {
+    Object.assign(styles, {borderTopRightRadius: rtr})
+  }
+  if (rbl) {
+    Object.assign(styles, {borderBottomLeftRadius: rbl})
+  }
+  if (rbr) {
+    Object.assign(styles, {borderBottomRightRadius: rbr})
   }
   return styles
 }
@@ -172,60 +211,5 @@ const spacing = (theme: BaseTheme, propValue: Spacing) => {
       return theme.spacing.space4
     default:
       return theme.spacing.space0
-  }
-}
-
-const lookupElevation = (theme: BaseTheme, elevationValue: number) => {
-  if (theme === undefined) {
-    throw Error(THEME_NOT_AVAILABLE_ERR_MSG)
-  }
-  switch (elevationValue) {
-    case 1: {
-      return theme.shadows.elevation1
-    }
-    case 2: {
-      return theme.shadows.elevation2
-    }
-    case 3: {
-      return theme.shadows.elevation3
-    }
-    case 4: {
-      return theme.shadows.elevation4
-    }
-    case 5: {
-      return theme.shadows.elevation5
-    }
-    case 6: {
-      return theme.shadows.elevation6
-    }
-    case 7: {
-      return theme.shadows.elevation7
-    }
-    case 8: {
-      return theme.shadows.elevation8
-    }
-    case 9: {
-      return theme.shadows.elevation9
-    }
-    case 10: {
-      return theme.shadows.elevation10
-    }
-    case 11: {
-      return theme.shadows.elevation11
-    }
-    case 12: {
-      return theme.shadows.elevation12
-    }
-    case 13: {
-      return theme.shadows.elevation13
-    }
-    case 14: {
-      return theme.shadows.elevation14
-    }
-    case 15: {
-      return theme.shadows.elevation15
-    }
-    default:
-      return theme.shadows.elevation0
   }
 }